requirements.txt
streamlit>=1.36,<2
plotly>=5.0,<6
pandas>=2.0,<3
requests>=2.31
openai>=1.30
apscheduler>=3.10,<4
cryptography>=42.0
py-clob-client>=0.1
python-dotenv>=1.0
